Minor changes in libraries

+++ src/logger/logger.c
@@ -4,11 +4,11 @@
 #include <stddef.h>
 #include <stdio.h>
 
-static size_t calculate_buffer_size(const char* format, va_list args) {
+static int calculate_buffer_size(const char* format, va_list args) {
     va_list args_copy;
     va_copy(args_copy, args);
 
-    const size_t size = vsnprintf(NULL, 0, format, args_copy) + 1;
+    const int size = vsnprintf(NULL, 0, format, args_copy) + 1;
 
     va_end(args_copy);
     return size;
@@ -19,7 +19,7 @@ void flb_logger(const char* category, const char* function, const char* format, 
         va_list args;
         va_start(args, format);
 
-        const size_t buffer_size = calculate_buffer_size(format, args);
+        const int buffer_size = calculate_buffer_size(format, args);
         char buffer[buffer_size];
 
         vsnprintf(buffer, buffer_size, format, args);
